To recapture the goal of this exercise, I think the primary goal was that we wanted to make sure that the RI can run in the mustang-like environment (IOW the RI in the system class loader), so that we can find  bugs before a PIT testing.
The secondary goal is to make sure that we can replace the RI in Mustang with a relatively simple procedure (namely to put jaxb-impl.jar into the bootstrap classpath or the endorsed directory.)

Somewhat related to this, the RI already have a mode of testing where it takes an arbitrary 1.0.x version of the JAXB RI, and run it on a mustang-like environment, and make sure that 1.0.x app will keep running.


I think it helps to think of the 1st and 2nd goals separately. Testing the 1st goal actually doesn't require JDK 6.0; we just need to set up a class loader that mimics Mustang-like environment.

The 2nd goal requires Mustang, but I think this goal can be achieved simply by running the same above configuration in the Mustang VM.


After thinking about this more, it seems to me that the test doesn't need to worry about using two VMs. If we run everything on Tiger, we can test the 1st goal. If we also impose on ourselves to be able to build on Mustang (not just run on Mustang), we can test the 2nd goal simply by running everything on Mustang.

This seems to simplify the build script quite a bit. You won't need the ${jvm} setting, nor do you need to play with two different versions of tools.jar. You just need to have a mode of class loader configuration where we put the runtime into the bootstrap classpath.

We haven't attempted to build the JAXB RI on Mustang, so there we might need some more work, but I think that effort won't be wasted, as we'll eventually have be able to build on Mustang anyway.

I made changes to the build script along these lines.