fix for bugid 6324461: JAXBContext.generateSchema(SchemaOutputResolver outputResolver) always throws Exception
The JAXBContextImpl.generateSchema() api wasn't overriding the javax.xml.bind version of the method
because there was a difference in which version of SchemaOutputResolver was being used.  I've removed
the com.sun.xml.bind.api version and switched all of the code over to using the javax.xml.bind version.