marshaller performance improvement.
We used to use CharSequence as the single data representation for all "PCDATA" that gets written. This was nice in the sense that I only needed to have a single path for various marshalling.

Now, as concrete instances, we use two kinds of different PCDATA representations. One is "java.lang.String", which is by far the most common datatype used in a schema, and the other is a custom PCDATA representation (such as Base64Data) that lazily turns itself into a String.

We used to have the single print routine for both, where we use CharSequence.length() and CharSequence.charAt() to examine one char at a time to write it out. This doesn't work well for many custom datatypes, such as "int". What we wanted was a way to let the PCDATA object itself to write it out. In this change this concept is realized as the Pcdata class. This design allows those custom PCDATA classes to skip the XML escape altogether, if it knows that it's not necessary.

With this change, we now needed to handle CharSequence in two different ways. One is for String where XmlOutput pulls data from it and send it to the stream; the other is for Pcdata, where XmlOutput lets it send itself to the stream. This check can be done by using "instanceof", but given that String is the majority, I felt that this is somewhat unfortunate. The producer of CharSequence knows whether it's a String or Pcdata, so it should be possible to just propage this information to XmlSerializer so that this "instanceof" can be avoided.

This thought leads to this change, where XMLSerializer now has two kinds of methods; one that takes String, the other that takes Pcdata.

I guess on reflection it wasn't too clear if this justifies adding this much code, but we'll see.